Added typecast to int to ensure compatibility with RPIO library on RasPi

+++ controlServos.py
@@ -18,11 +18,11 @@ def steer(deg):
 		return
 	
 	elif (deg < 0):
-		servo1.set_servo(17, BASE + (-deg)*LEFT)
+		servo1.set_servo(17, int(BASE + (-deg)*LEFT))
 		servo1.stop_servo(17)
 		
 	elif (deg > 0):
-		servo1.set_servo(17, BASE + deg*RIGHT)
+		servo1.set_servo(17, int(BASE + deg*RIGHT))
 		servo1.stop_servo(17)
 		
 	elif (deg == 0):
@@ -36,11 +36,11 @@ def drive(speed):
 		return
 		
 	elif (speed < 0):
-		servo2.set_servo(22, BASE + (-speed)*BACK)
+		servo2.set_servo(22, int(BASE + (-speed)*BACK))
 		servo2.stop_servo(22)		
 	
 	elif (speed > 0):
-		servo2.set_servo(22, BASE + speed*FORWARD)
+		servo2.set_servo(22, int(BASE + speed*FORWARD))
 		servo2.stop_servo(22)
 	
 	elif (speed == 0):
